improve and rename MonthlySummary class
